update#1.05
The toolbox is ready to import ECG-PDFs (*.pdf) from Apple Watch and AliveCor Kardia that works on all main OS (MacOS, Windows, Linux/UNIX). It also imports the beat ticks (annotations) that comes with the PDF. All you need is a free installation of Inkscape and you need to start Matlab with administrator rights. For Linux systems a faster solution exist when having the packages pdf2svg and pdfminer installed. For the import of PDFs in batch mode you can use the function read_pdf that converts the pdf to Matlab structure. It's now possible to save the annotations as EDF files and to import signals and annotations from workspace variables. Therefore you need to cancel the import of a file, that automatically directs you to a selection dialog of workspace variables. This is also working when loading annotations. Moreover, the power spectrum was not squared, which is now corrected. The table of HRV parameters contains no Approximate Entropy ApEn.